// libs/core/src/identifier-mapping/infrastructure/persistence/entities/identifier-mapping.orm-entity.d.ts
export declare class IdentifierMappingOrmEntity {
    id: string;
    entityType: string;
    internalId: string;
    externalId: string;
    platformType: string;
    connectionId: string;
    context: Record<string, unknown> | null;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=identifier-mapping.orm-entity.d.ts.map